+++ SRC/NAVIGATION/ahrs.c
@@ -101,7 +101,7 @@ void AttitudeEstimate(Vector3f_t gyro, Vector3f_t acc, Vector3f_t mag)
 {
     Vector3f_t deltaAngle;	
     Vector3f_t accCompensate;
-	static uint32_t previousT;
+	static uint64_t previousT;
 	float deltaT = (GetSysTimeUs() - previousT) * 1e-6;	
     deltaT = ConstrainFloat(deltaT, 0.0005, 0.002);	
 	previousT = GetSysTimeUs();
